arreglos en kardex

+++ src/main/java/com/inventario/models/SalidaDetalleInventario.java
@@ -146,7 +146,7 @@ public class SalidaDetalleInventario {
         return subtotal;
     }
 
-    public void setSubtotal(int subtotal) {
+    public void setSubtotal(double subtotal) {
         this.subtotal = subtotal;
     }
